I decided that there is no point in using 'long' to store offsets to transport specification file -- if that file exceeds 2G in 32-bit systems we are in deep trouble anyway -- and there is no point in trying to fit this animal into 16-bit machines...  Could squeeze around 16 bytes per recipient in the scheduler memory use. (Consider having a million recipients in memory data...)

+++ include/ta.h
@@ -187,7 +187,7 @@ extern int erename __((char *from, char *to));
 #endif
 
 /* lockaddr.c: */
-extern int lockaddr __((int fd, char *map, long offset, int was, int new, const char *file, const char *host, const int mypid));
+extern int lockaddr __((int fd, char *map, int offset, int was, int new, const char *file, const char *host, const int mypid));
 
 /* markoff.c: */
 extern int markoff __((char *filecontents, int bytesleft, long offsets[], const char *filename));
